Reformat using the provided clang-format configuration
This is to simplify subsequent commits. Mostly unifies white space, reorders
a few includes and attaches pointer '*' to the type instead of the variable.

The most visible change is forcing return type on a separate line, which
cannot be configured to be left on the user.

Clang-format can finally deal with custom bin packing of function parameters
and arguments and leaves it as is. Split from "Tweak comment in build.xml"