Simplify dual-loading support
Indicating that the agent has been already loaded by setting a field
in the 'UbenchAgent' class can fail in an environment with multiple
class loaders, because we may fail to find the 'UbenchAgent' class.

Instead, we just check if we can call the agent's native functions and
if not, we try to load it explicitly using the 'loadLibrary()' method.